Minor refactor: reorganize layout
